+++ beem/__init__.py
@@ -1,9 +1,11 @@
 """ beem."""
 from .steem import Steem
 from .hive import Hive
+from .blurt import Blurt
 from .version import version as __version__
 __all__ = [
     "steem",
+    "blurt",
     "account",
     "amount",
     "asset",