+++ beem/asset.py
@@ -48,7 +48,7 @@ class Asset(BlockchainObject):
         self.chain_params = self.blockchain.get_network()
         if self.chain_params is None:
             from beemgraphenebase.chains import known_chains
-            self.chain_params = known_chains["STEEMAPPBASE"]
+            self.chain_params = known_chains["STEEM"]
         self["asset"] = ""
         found_asset = False
         for asset in self.chain_params["chain_assets"]: